+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Support for loading cookie files from netscape cookie files.
+- Support specifying the HTTP request interval with `--interval`
 
 ## [0.1.0] - 2019-01-20